Fixes for Audio commands not replying
- Refactored so that every message is deferred and replied to later
unless there is no operations being done. e.g., when playing a track,
defers the response just in case the track adding takes more than 3
seconds (the limit of responding to a slash command) and then follows up
 with the actual response when it is finished.
- Should fix issues around messages not coming through or throwing
errors without the user knowing why its not happening
- Added cases for where the user is in a voice channel but the bot
isn't and vice versa to inform the user what the issue is with the
command they just tried to execute and why it failed.
- Renamed "offset" to "page" inside the queue command to make more sense
 to the user